refactor(agent): extract tool execution boundary
